Add integration tests for refunds, subscriptions, tenants, and tests services
- Implement integration tests for the refunds service to validate tenant ownership, payment status, and provider payment ID checks.
- Create integration tests for the subscriptions service to cover checkout, verification, and cancellation flows.
- Add a regression test for the tenants service to ensure domain registration checks are functioning correctly.
- Introduce a comprehensive integration test for the tests service to verify the full test-taking and auto-scoring pipeline.
- Create a new database role for the application to enforce row-level security and prevent superuser access.
- Implement scripts for nightly database backups and restoration from backups, ensuring data safety and recoverability.

+++ sql/queries/enrollments.sql
@@ -26,10 +26,14 @@ ORDER BY e.enrolled_at DESC
 LIMIT $2 OFFSET $3;
 
 -- name: UpdateEnrollmentProgress :exec
+-- $2 cast to numeric explicitly — without it, pgx's parameter-type
+-- inference sees `$2 >= 100` (untyped literal) and `progress_percent =
+-- $2` (numeric(5,2) column) and can't unify a single type for $2 across
+-- both usages, failing with "inconsistent types deduced for parameter".
 UPDATE enrollments
-SET progress_percent = $2,
-    status = CASE WHEN $2 >= 100 THEN 'completed' ELSE status END,
-    completed_at = CASE WHEN $2 >= 100 THEN CURRENT_TIMESTAMP ELSE completed_at END
+SET progress_percent = $2::numeric,
+    status = CASE WHEN $2::numeric >= 100 THEN 'completed' ELSE status END,
+    completed_at = CASE WHEN $2::numeric >= 100 THEN CURRENT_TIMESTAMP ELSE completed_at END
 WHERE id = $1;
 
 -- name: CancelEnrollment :exec
